refactor(engine): retire runtime PLY parser (M1 close / #396)
Sever the runtime PLY parser from gseurat_core. After this PR, the engine
binary cannot fall back to text-PLY parsing — every cloud load must hit a
sibling .gsvx baked by PR-1's bake_gsvx target.

Changes:

- Extract load_ply / write_ply and their helpers (PlyProperty, type_size,
  read_float, kSH_C0) from src/engine/gaussian_cloud.cpp into a new
  src/engine/gaussian_cloud_ply.cpp. The new file is NOT compiled into
  gseurat_core — only into ply2heightmap and tests that exercise PLY
  fixtures.
- Tighten GaussianCloud::load_with_gsvx_first to throw std::runtime_error
  when the sibling .gsvx is missing, unreadable, or fails to parse. No
  more silent fallback; PR-1 guarantees coverage for bundled assets, and
  anything missing now is a build/asset-pipeline error.
- Update header doc-comments on load_ply (offline-only) and
  load_with_gsvx_first (no fallback) to match the new contract.
- CMake: every place that listed gaussian_cloud.cpp outside gseurat_core
  (ply2heightmap + 10 test targets) now also lists gaussian_cloud_ply.cpp.
  The gseurat_core source list is untouched.
- Invert test_gaussian_cloud Tests 24, 26, 27: they previously asserted
  fallback behavior (missing/corrupt/truncated sibling) and now assert
  the wrapper throws. Tests 25, 28, 29 (GSVX-first path with sibling
  present) unchanged.

Validation:
- cmake --build (demo, staging, ply2heightmap) clean.
- test_gaussian_cloud all 29 tests pass, including the 3 inverted.
- 10s runtime smoke: demo loads seurat_island.ply scene, 7 BoneAnimated
  characters, dozens of PBD trees — zero throws, zero missing-sibling
  errors. Every cloud load served from GSVX.

This closes M1 (Phase 1 end) of #396. Phase 2 (instrumentation:
GS_LABEL per dispatch, GS_DBG_INVARIANT migration) is next per
docs/superpowers/specs/2026-05-02-engine-refactor-phase1-design.md §6.
