+++ tests/test_gaussian_cloud.cpp
@@ -821,24 +821,22 @@ int main() {
         write_test_gsvx_v2(gsvx_path, data, aabb);
     };
 
-    // ====== Test 24: fallback to PLY when no .gsvx sibling exists ======
+    // ====== Test 24: throws when no .gsvx sibling exists ======
+    // Post-#396 M1: there is no PLY fallback in the runtime path. The
+    // bake_gsvx CMake target guarantees a sibling .gsvx for every bundled
+    // .ply; a missing one is a build/asset-pipeline error.
     {
-        const std::string ply_path = tmp_dir + "/test_dual_fallback.ply";
+        const std::string ply_path = tmp_dir + "/test_dual_no_sibling.ply";
         write_test_ply(ply_path, 5);
 
-        auto via_ply  = GaussianCloud::load_ply(ply_path);
-        auto via_dual = GaussianCloud::load_with_gsvx_first(ply_path);
-
-        assert(via_dual.count() == via_ply.count() && "fallback count matches");
-        for (uint32_t i = 0; i < via_ply.count(); ++i) {
-            const auto& a = via_ply.gaussians()[i];
-            const auto& b = via_dual.gaussians()[i];
-            assert(approx(a.position.x, b.position.x, 1e-6f) && "fallback position.x");
-            assert(approx(a.opacity,    b.opacity,    1e-6f) && "fallback opacity");
-            assert(approx(a.scale.x,    b.scale.x,    1e-6f) && "fallback scale.x");
-            assert(approx(a.importance, b.importance, 1e-6f) && "fallback importance");
+        bool threw = false;
+        try {
+            (void)GaussianCloud::load_with_gsvx_first(ply_path);
+        } catch (const std::runtime_error&) {
+            threw = true;
         }
-        printf("PASS: Test 24 - load_with_gsvx_first falls back to PLY when no sibling\n");
+        assert(threw && "load_with_gsvx_first throws when no sibling .gsvx");
+        printf("PASS: Test 24 - load_with_gsvx_first throws when no sibling .gsvx\n");
     }
 
     // ====== Test 25: prefers .gsvx sibling when present (parity with load_ply) ======
@@ -882,12 +880,12 @@ int main() {
         printf("PASS: Test 25 - load_with_gsvx_first prefers sibling .gsvx (parity)\n");
     }
 
-    // ====== Test 26: graceful fallback to PLY when .gsvx is corrupt ======
+    // ====== Test 26: throws when sibling .gsvx is corrupt ======
     {
         const std::string ply_path  = tmp_dir + "/test_dual_corrupt.ply";
         const std::string gsvx_path = tmp_dir + "/test_dual_corrupt.gsvx";
         write_test_ply(ply_path, 4);
-        // Write a bad-magic GSVX next to the PLY
+        // Write a bad-magic GSVX next to the PLY.
         {
             std::ofstream out(gsvx_path, std::ios::binary);
             GsvxHeader header{};
@@ -896,25 +894,23 @@ int main() {
             out.write(reinterpret_cast<const char*>(&header), sizeof(header));
         }
 
-        auto via_ply  = GaussianCloud::load_ply(ply_path);
-        auto via_dual = GaussianCloud::load_with_gsvx_first(ply_path);
-
-        assert(via_dual.count() == via_ply.count() && "corrupt fallback count");
-        for (uint32_t i = 0; i < via_ply.count(); ++i) {
-            assert(approx(via_ply.gaussians()[i].position.x,
-                          via_dual.gaussians()[i].position.x, 1e-6f) &&
-                   "corrupt fallback position.x");
+        bool threw = false;
+        try {
+            (void)GaussianCloud::load_with_gsvx_first(ply_path);
+        } catch (const std::runtime_error&) {
+            threw = true;
         }
-        printf("PASS: Test 26 - load_with_gsvx_first falls back on corrupt sibling\n");
+        assert(threw && "load_with_gsvx_first throws on corrupt sibling .gsvx");
+        printf("PASS: Test 26 - load_with_gsvx_first throws on corrupt sibling\n");
     }
 
-    // ====== Test 27: graceful fallback when sibling .gsvx is truncated ======
+    // ====== Test 27: throws when sibling .gsvx is truncated ======
     {
         const std::string ply_path  = tmp_dir + "/test_dual_short.ply";
         const std::string gsvx_path = tmp_dir + "/test_dual_short.gsvx";
         write_test_ply(ply_path, 3);
-        // Write a header that claims count=999 but contains no payload — load_gsvx
-        // throws on size mismatch; wrapper must catch and fall through.
+        // Header claims count=999 but contains no payload — load_gsvx
+        // throws on size mismatch and the wrapper now propagates it.
         {
             std::ofstream out(gsvx_path, std::ios::binary);
             GsvxHeader header{};
@@ -925,11 +921,14 @@ int main() {
             out.write(reinterpret_cast<const char*>(&header), sizeof(header));
         }
 
-        auto via_ply  = GaussianCloud::load_ply(ply_path);
-        auto via_dual = GaussianCloud::load_with_gsvx_first(ply_path);
-
-        assert(via_dual.count() == via_ply.count() && "truncated fallback count");
-        printf("PASS: Test 27 - load_with_gsvx_first falls back on truncated sibling\n");
+        bool threw = false;
+        try {
+            (void)GaussianCloud::load_with_gsvx_first(ply_path);
+        } catch (const std::runtime_error&) {
+            threw = true;
+        }
+        assert(threw && "load_with_gsvx_first throws on truncated sibling .gsvx");
+        printf("PASS: Test 27 - load_with_gsvx_first throws on truncated sibling\n");
     }
 
     // ====== Test 28: kVulkanYDown applied symmetrically on GSVX path ======
